Jan pointed out some issues with CPE 2.2 vs 2.3
It seems that CPE, which is supposed to be magic and figure out which
version of CPE its seeing on its own, has some slight issues and will
validate a 2.3 URI using the packed method as valid for 2.2 (which it
isn't) so take his suggestion to use CPE2_2 further and add a --mode so
you can validate with either, and add some other cosmetic niceties so
you don't see useless information (i.e. see 2.3 things for a 2.2 CPE,
etc.).